đánh giá sản phẩm

+++ be/src/main/java/com/project/pharmacy/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findById(int id);
     Optional<User> findByEmailAndPassword(String email, String password);
     User findByName(String username);
@@ -13,5 +13,4 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     User findUserById(int id);
     Optional<User> findByEmail(String email);
-
 }
